fix frontend problmes

+++ backend/app/schemas/dashboard.py
@@ -1,6 +1,5 @@
 import uuid
 from datetime import date
-from decimal import Decimal
 
 from pydantic import BaseModel
 
@@ -8,20 +7,20 @@ from pydantic import BaseModel
 class WeekProjection(BaseModel):
     week_id: uuid.UUID | None
     week_start: date
-    projected_balance: Decimal
+    projected_balance: float
 
 
 class GoalDelta(BaseModel):
     id: uuid.UUID
     name: str
-    target_amount: Decimal
-    current_amount: Decimal
-    remaining: Decimal
+    target_amount: float
+    current_amount: float
+    remaining: float
     target_date: date
     status: str
 
 
 class DashboardSummary(BaseModel):
-    current_balance: Decimal
+    current_balance: float
     projection: list[WeekProjection]
     goals: list[GoalDelta]
